Move HTML cache to new class and update tests
