+++ src/test/core/test_core.py
@@ -18,6 +18,7 @@ from glob import glob
 from string import Template
 
 from djerba.core.configure import config_wrapper, core_configurer, DjerbaConfigError
+from djerba.core.html_cache import html_cache, DjerbaHtmlCacheError
 from djerba.core.ini_generator import ini_generator
 from djerba.core.json_validator import plugin_json_validator
 from djerba.core.loaders import plugin_loader, core_config_loader, DjerbaLoadError
@@ -384,16 +385,17 @@ class TestHtmlCache(TestCore):
     """Test the HTML cache"""
 
     def test_encode_decode(self):
+        # test an encoding/decoding round trip
         string_to_encode = "Hello, world!"
-        djerba_main = main(self.tmp_dir, log_level=logging.ERROR)
-        encoded = djerba_main.encode_to_base64(string_to_encode)
-        decoded_string = djerba_main.decode_from_base64(encoded)
+        cache = html_cache(log_level=logging.ERROR)
+        encoded = cache.encode_to_base64(string_to_encode)
+        decoded_string = cache.decode_from_base64(encoded)
         self.assertEqual(string_to_encode, decoded_string)
 
     def test_html_wrap(self):
-        djerba_main = main(self.tmp_dir, log_level=logging.ERROR) # suppress author warning
+        cache = html_cache(log_level=logging.ERROR)
         html_string = "<p>Hello, world!</p>"
-        wrapped = djerba_main.wrap_html('test', html_string)
+        wrapped = cache.wrap_html('test', html_string)
         lines = [
             "<span DJERBA_COMPONENT_START='test' />",
             "<p>Hello, world!</p>",
@@ -409,30 +411,30 @@ class TestHtmlCache(TestCore):
         broken2 = "blahblah<span DJERBA_COMPONENT_START='test'/>"
         broken3 = "lorem ipsum dolor"
         broken4 = "<span DJERBA_COMPONENT_NOWHERE='test'/>"
-        djerba_main = main(self.tmp_dir, log_level=logging.CRITICAL)
-        self.assertEqual(djerba_main.parse_name_from_separator(good), 'test')
+        cache = html_cache(log_level=logging.CRITICAL)
+        self.assertEqual(cache.parse_name_from_separator(good), 'test')
         for tag in [broken1, broken2, broken3, broken4]:
             with self.assertRaises(DjerbaHtmlCacheError):
-                djerba_main.parse_name_from_separator(tag)
+                cache.parse_name_from_separator(tag)
 
     def test_update_cached(self):
         with open(os.path.join(self.test_source_dir, 'example.html')) as in_file:
             old_html_string = in_file.read()
         tmp_log = os.path.join(self.tmp_dir, 'djerba.log')
-        djerba_main = main(self.tmp_dir, log_level=logging.DEBUG, log_path=tmp_log)
-        old_html = djerba_main.encode_to_base64(old_html_string)
+        cache = html_cache(log_level=logging.DEBUG, log_path=tmp_log)
+        old_html = cache.encode_to_base64(old_html_string)
         new_html = {
             'test': "<span DJERBA_COMPONENT_START='test'/>\n"+\
             'Duis aute irure dolor in reprehenderit in voluptate velit esse '+\
             'cillum dolore eu fugiat nulla pariatur. This is a Djerba test.\n'+\
             "<span DJERBA_COMPONENT_END='test'/>"
         }
-        with self.assertLogs('djerba.core.main', level=logging.DEBUG) as log_context:
-            updated_html = djerba_main.update_cached_html(new_html, old_html)
+        with self.assertLogs('djerba.core.html_cache', level=logging.DEBUG) as log_context:
+            updated_html = cache.update_cached_html(new_html, old_html)
         log_messages = [
-            'DEBUG:djerba.core.main:Updating test',
-            'DEBUG:djerba.core.main:No update found for test_no_update',
-            'DEBUG:djerba.core.main:HTML update done'
+            'DEBUG:djerba.core.html_cache:Updating test',
+            'DEBUG:djerba.core.html_cache:No update found for test_no_update',
+            'DEBUG:djerba.core.html_cache:HTML update done'
         ]
         for msg in log_messages:
             self.assertIn(msg, log_context.output)
@@ -440,14 +442,14 @@ class TestHtmlCache(TestCore):
         self.assertFalse('Ut enim ad minim veniam, quis nostrud' in updated_html)
         self.assertTrue('Excepteur sint occaecat cupidatat non proident' in updated_html)
         # test if the updated html can be updated again
-        old_html_2 = djerba_main.encode_to_base64(updated_html)
+        old_html_2 = cache.encode_to_base64(updated_html)
         new_html_2 = {
             'test': "<span DJERBA_COMPONENT_START='test2'/>\n"+\
             'Duis aute irure dolor in reprehenderit in voluptate velit esse '+\
             "cillum dolore eu fugiat nulla pariatur. This is another Djerba test.\n"+\
             "<span DJERBA_COMPONENT_END='test2'/>",
         }
-        updated_html_2 = djerba_main.update_cached_html(new_html_2, old_html_2)
+        updated_html_2 = cache.update_cached_html(new_html_2, old_html_2)
         self.assertTrue('This is another Djerba test.' in updated_html_2)
         self.assertFalse('This is a Djerba test.' in updated_html_2)
         self.assertTrue('Excepteur sint occaecat cupidatat non proident' in updated_html)
